+++ src/app/list-page/list-page.component.html
@@ -112,6 +112,9 @@
           >
             External metadata
           </ion-item>
+          <ion-item button detail="false" (click)="exportSelectedGamesFromPopover()">
+            Export CSV
+          </ion-item>
           <ion-item button detail="false" color="danger" (click)="deleteSelectedGames()">
             Delete
           </ion-item>
